Add delay after the slurmdbd.service is started.
The problem Alan Sill was facing was due to the fact that the Cluster
is not being registered with slurm database before sacctmgr started its business.
so sacctmgr is not able to load it's config file.
For the cluster name to be registered in time for the sacctmgr to operate we added
a 5 sec delay.