Add task to import /etc/passwd,group,shadow files.
This is necessary to avoid the no object found error, when we do the
wwsh file resync. Previously when we had the compute_build_nodes role
running on the ohpc node we did a file sync after we set the files
(passwd,shadow etc.) for provisioning with the nodes. SO they were
already in the WW DB hence file sync worked across all the roles but
now we moved it to post image build play book ohpc-ops.yaml so we add
the necessary files to WW DB, make changes as needed and then do a file
sync to update the DB.