disable firewalld to mount the home dir in OOD.
The home dir would not mount on OOD if we did not disable the firewalld
in OHPC, it was working in our previous setup while deploying in vagrant
as the VBox VM was running and was never turned off during the setup,
but in the cloud deploy we have to preserve the state we have in the
build phase (i.e. after packer builds the image). We need to disable
it so it would not get turned back on after we boot the image during
deploy