Change which RMQ agents are installed based on prod or dev deploy
For a prod deploy only agents inside prod_rmq_agents are deployed.
For a dev deploy only agents inside dev_rmq_agents are deployed.
The agent template agent.j2 requires the PYTHONPATH to be set for
agents to refer to rc_rmq and rc_util modules.
The working dir in the agent is set in accordance with dev/prod deploy.